Backend/requirements.txt
azure-core==1.35.1
azure-cosmos==4.9.0
azure-data-tables==12.5.0
azure-functions==1.24.0
azure-identity==1.25.1
azure-storage-blob==12.27.0
azure-storage-queue==12.14.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
idna==3.11
isodate==0.7.2
MarkupSafe==3.0.3
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.0
propcache==0.4.1
pycparser==2.23
PyJWT==2.10.1
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.22.0
